Fix auto-tools build and test.
test.sh.in: only the ipopt tests need test.sh.in.
rev_sparse_jac.cpp: avoid warning when using eigen vectors.

git-svn-id: https://projects.coin-or.org/svn/CppAD/branches/atomic@2840 0e562018-b8fb-0310-a933-ecf073c7c197

+++ test_more/rev_sparse_jac.cpp
@@ -422,7 +422,7 @@ bool case_four()
 	s[1] = false;
 	r = F.RevSparseJac(q, s);
 
-	ok &= r.size() == q * n;
+	ok &= size_t( r.size() ) == q * n;
 	ok &= r[0] == false;
 	ok &= r[1] == true;
